gateway/requirements.txt
fastapi
uvicorn[standard]
grpcio
protobuf
python-dotenv